PopupWindow : Replace with simpler implementation from PlugPopup
The old features weren't useful or reliable.
