+++ python/GafferSceneUI/LightEditor.py
@@ -269,10 +269,7 @@ class LightEditor( GafferUI.NodeSetEditor ) :
 
 		else :
 
-			# See todo in `PlugPopup._PopupWindow`
-			PopupWindow = GafferUI.PlugPopup.__bases__[0]
-
-			with PopupWindow() as self.__popup :
+			with GafferUI.PopupWindow() as self.__popup :
 				with GafferUI.ListContainer( GafferUI.ListContainer.Orientation.Horizontal, spacing = 4 ) :
 					GafferUI.Image( "warningSmall.png" )
 					GafferUI.Label( "<h4>{}</h4>".format( inspection.nonEditableReason() ) )
